# Project/backend/src/main/java/com/vem/backend/dto/RegisterDto.java
package com.vem.backend.dto;


public class RegisterDto {
    private String name;
    private String email;
    private String password;
    private String phone;

    public RegisterDto() {
    }

    public RegisterDto(String name, String email, String password, String phone) {
        this.name = name;
        this.email = email;
        this.password = password;
        this.phone = phone;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public String getEmail() {
        return email;
    }

    public void setEmail(String email) {
        this.email = email;
    }

    public String getPassword() {
        return password;
    }

    public void setPassword(String password) {
        this.password = password;
    }

    public String getPhone() {
        return phone;
    }

    public void setPhone(String phone) {
        this.phone = phone;
    }

}